Split utils in two
The new `search_util.rb` is exclusively for those methods that are for
navigating the store.
Closes #49
