+++ lib/inventoryware.rb
@@ -40,6 +40,7 @@ require_rel 'command'
 require_rel 'exceptions'
 require_rel 'commands/**/*.rb'
 require_rel 'lsblk_parser'
+require_rel 'search_utils'
 require_rel 'utils'
 
 require 'commander/import'
